speed

+++ commands/music/pitch.js
@@ -8,7 +8,7 @@ class PitchCommand extends MusicCommand {
     if (!this.guild.voiceStates.get(this.client.user.id)?.channelID) return this.getString("sound.notInVoice");
     if (!this.connection) return this.getString("sound.noConnection");
     if (this.connection.host !== this.author.id && !this.memberPermissions.has("MANAGE_CHANNELS")) return "Only the current voice session host can change the pitch!";
-    const vol = Number.parseInt(this.options.level ?? this.args[0]);
+    const vol = Number.parseFloat(this.options.level ?? this.args[0]);
     if (Number.isNaN(vol) || vol > 1000 || vol < 0) return "You can only set the pitch between 0 and 1000!";
     const settings = {
       pitch: vol
